refactor: Repository naming 수정
- 향후 Mybatis, Jpa 사용 고려하여 구체적인 데이터 접근 기술 명시
