+++ src/main/java/com/flab/livecommerce/infrastructure/item/persistence/jdbctemplate/JdbcTemplateItemOptionRepository.java
@@ -1,4 +1,4 @@
-package com.flab.livecommerce.infrastructure.item.persistence.mysql;
+package com.flab.livecommerce.infrastructure.item.persistence.jdbctemplate;
 
 import com.flab.livecommerce.domain.item.ItemOption;
 import javax.sql.DataSource;
@@ -9,12 +9,12 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class MysqlItemOptionRepository {
+public class JdbcTemplateItemOptionRepository {
 
     private JdbcTemplate template;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public MysqlItemOptionRepository(DataSource dataSource) {
+    public JdbcTemplateItemOptionRepository(DataSource dataSource) {
         this.template = new JdbcTemplate(dataSource);
         this.jdbcInsert = new SimpleJdbcInsert(dataSource)
             .withTableName("item_option")
